added shader params, base compute pass class
